+++ src/shared/helpers/environment.ts
